move TodoDatabase in dao package
