+++ app/src/main/java/com/areebgroup/todolist/ui/MainActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.material3.Surface
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.areebgroup.todolist.ui.theme.ToDoListTheme
 import com.areebgroup.todolist.ui.todo.view.HomeScreen
 import dagger.hilt.android.AndroidEntryPoint
