+++ app/src/main/java/com/areebgroup/todolist/ui/todo/view/components/ContentSection.kt
@@ -40,7 +40,7 @@ fun ContentSection(
                 modifier = Modifier
                     .padding(start = 8.dp)
                     .size(70.dp),
-                painter = painterResource(id = R.drawable.icon_image),
+                painter = painterResource(id = R.drawable.app_icon),
                 contentDescription = "App icon"
             )
